moved CardModal from Card.js into its own component folder and file
