+++ server/client/src/features/cardModal/CardModal.module.css
@@ -1,0 +1,34 @@
+.cardLabel {
+  margin: 9px;
+  margin-top: 11px;
+  height: 25px;
+  width: 60px;
+}
+
+.list {
+  font-weight: 700;
+  text-decoration: underline;
+}
+
+.selectBtn {
+  cursor: pointer;
+}
+
+.selectBtn:hover {
+  background-color: rgba(173, 200, 210, 0.5);
+}
+
+.actionButton {
+  cursor: pointer;
+  margin-bottom: 5px;
+  height: 38px;
+  padding-left: 10px;
+  padding-top: 6px;
+  border: 1px solid #dee2e6;
+  border-radius: 5px;
+  background-color: rgba(173, 200, 210, 0.5);
+}
+
+.actionButton:hover {
+  background-color: rgba(173, 200, 210, 0.8);
+}
